Onboarding page is divided on different files
