+++ constants/Stocks.ts
@@ -1,0 +1,4 @@
+export default class StockModel {
+    title: string = "";
+    isSelected: boolean = false;
+}
